(doprnt1): Cast arg to CHAR_STRING.

+++ src/doprnt.c
@@ -293,7 +293,7 @@ doprnt1 (lispstrings, buffer, bufsize, format, format_end, nargs, args)
 	    case 'c':
 	      if (cnt == nargs)
 		error ("not enough arguments for format string");
-	      tem = CHAR_STRING ((EMACS_INT) args[cnt], charbuf, string);
+	      tem = CHAR_STRING ((int) (EMACS_INT) args[cnt], charbuf, string);
 	      cnt++;
 	      string[tem] = 0;
 	      width = strwidth (string, tem);
